refactor(graphics): split camera projection module (Issue #284 Phase2)
