+++ viewmodel/graphics/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod camera;
 mod camera_math;
+mod camera_projection;
 
 pub use camera::{build_view_projection_matrix, Camera, CameraControlSensitivity};
